Fix some of unused vars build
