+++ src/actions/ticket.ts
@@ -1,9 +1,8 @@
 'use server'
 
-import { authConfig } from "@/app/api/auth/[...nextauth]/route";
 import { CreateTicketFormSchema, NotificationWithUserAndContainer } from "@/lib/definitions";
 import { prisma } from "@/lib/prisma";
-import { isAdmin } from "@/lib/utils";
+import { authConfig, isAdmin } from "@/lib/utils";
 import { NotificationState, NotificationType } from "@prisma/client";
 import { getServerSession } from "next-auth";
 import { revalidatePath } from "next/cache";
