+++ src/components/dialogs/containers/EditMemoryLimitContainer.tsx
@@ -1,10 +1,10 @@
-import { editAdminCpuLimit, editAdminMemoryLimit } from "@/actions/containers";
+import { editAdminMemoryLimit } from "@/actions/containers";
 import { Button } from "@/components/ui/button";
 import { Dialog, DialogClose, DialogContent, DialogDescription, DialogFooter, DialogHeader, DialogTitle, DialogTrigger } from "@/components/ui/dialog";
 import { Form, FormControl, FormDescription, FormField, FormItem, FormLabel, FormMessage } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
 import { toast } from "@/hooks/use-toast";
-import { EditCpuLimitContainerFormSchema, EditMemoryLimitContainerFormSchema } from "@/lib/definitions";
+import { EditMemoryLimitContainerFormSchema } from "@/lib/definitions";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { Container } from "@prisma/client";
 import { Loader2 } from "lucide-react";
