+++ src/i18n/request.ts
@@ -6,7 +6,7 @@ export default getRequestConfig(async ({requestLocale}) => {
   let locale = await requestLocale;
  
   // Ensure that a valid locale is used
-  if (!locale || !routing.locales.includes(locale as any)) {
+  if (!locale || !routing.locales.includes(locale as "en" | "fr")) {
     locale = routing.defaultLocale;
   }
