+++ src/ui/fonts.ts
@@ -1,4 +1,4 @@
-import { Inter, Roboto, Roboto_Mono } from 'next/font/google';
+import { Inter, Roboto_Mono } from 'next/font/google';
 
 export const roboto = Inter({ subsets: ['latin'] })
 export const robotoMono = Roboto_Mono({ subsets: ['latin'] });
